app/user_service/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
django-extensions
gunicorn
psycopg2-binary
dj-database-url
django-environ
whitenoise
django-cors-headers
requests
httpx
pillow